divided into seperate files

+++ tasks/count-comparer.hpp
@@ -2,8 +2,8 @@
 // Function to count comparisons for std::priority_queue
 template <typename T>
 struct compare_count {
-    int& count;
-    compare_count(int& count) : count(count) {}
+    uint64_t& count;
+    compare_count(uint64_t& count) : count(count) {}
     bool operator()(const T& a, const T& b) {
         count++;
         return a > b;  // sorts in ascending order
